refactor(chat): 웹소켓 채팅 개선 #216
