+++ src/main/java/com/my/relink/chat/handler/StompHandler.java
@@ -2,6 +2,7 @@ package com.my.relink.chat.handler;
 
 import com.my.relink.chat.config.ChatPrincipal;
 import com.my.relink.chat.config.WebSocketSessionManager;
+import com.my.relink.chat.handler.metric.OperationMetrics;
 import com.my.relink.config.security.AuthUser;
 import com.my.relink.config.security.jwt.JwtProvider;
 import com.my.relink.domain.trade.Trade;
@@ -31,6 +32,8 @@ import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
+
+import static com.my.relink.chat.handler.metric.OperationMetrics.*;
 
 @Component
 @RequiredArgsConstructor
@@ -225,63 +228,5 @@ public class StompHandler implements ChannelInterceptor {
         return Long.parseLong(paths[paths.length - 2]);
     }
 
-    @Getter
-    public static class MetricsDTO {
-        long count;
-        double averageLatency;
-        long maxLatency;
-        long percentile95;
-
-        public MetricsDTO(long count, double averageLatency, long maxLatency, long percentile95) {
-            this.count = count;
-            this.averageLatency = averageLatency;
-            this.maxLatency = maxLatency;
-            this.percentile95 = percentile95;
-        }
-    }
-
-    private static class OperationMetrics {
-        private final AtomicLong count = new AtomicLong(0);
-        private final AtomicLong totalLatency = new AtomicLong(0);
-        private final AtomicLong maxLatency = new AtomicLong(0);
-        private final ConcurrentSkipListSet<Long> latencies = new ConcurrentSkipListSet<>();
-
-        public OperationMetrics addLatency(long latency) {
-            count.incrementAndGet();
-            totalLatency.addAndGet(latency);
-            updateMaxLatency(latency);
-            latencies.add(latency);
-            return this;
-        }
-
-        private void updateMaxLatency(long latency) {
-            long currentMax;
-            do {
-                currentMax = maxLatency.get();
-                if (latency <= currentMax) break;
-            } while (!maxLatency.compareAndSet(currentMax, latency));
-        }
-
-        public long getCount() {
-            return count.get();
-        }
-
-        public double getAverageLatency() {
-            long currentCount = count.get();
-            return currentCount > 0 ?
-                    (double) totalLatency.get() / currentCount : 0;
-        }
-
-        public long getMaxLatency() {
-            return maxLatency.get();
-        }
-
-        public long get95thPercentile() {
-            List<Long> sortedLatencies = new ArrayList<>(latencies);
-            if (sortedLatencies.isEmpty()) return 0;
-            int index = (int) Math.ceil(sortedLatencies.size() * 0.95) - 1;
-            return sortedLatencies.get(Math.max(0, Math.min(index, sortedLatencies.size() - 1)));
-        }
-    }
 
 }
